changed integer creation to long
git-svn-id: http://svn.hive.pt/hive-main/pt.hive.colony/trunk/pt.hive.colony@632 89aa9a68-8201-4d56-8cc5-2cc230aca7a0

+++ src/colony/plugins/util.py
@@ -158,7 +158,7 @@ def get_timestamp_uid():
     float_value = timestamp * (10 ** UID_PRECISION)
 
     # retrieves the integer value
-    integer_value = int(float_value)
+    integer_value = long(float_value)
 
     # converts the integer value to string value
     string_value = str(integer_value)
